requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.10.0
google-generativeai>=0.3.0
wikipedia>=1.4.0
ipython>=8.12.0
python-dotenv>=1.0.0
PyAudio>=0.2.11
Flask>=3.0.0
Flask-SocketIO>=5.3.0
eventlet>=0.33.0
psutil>=5.9.0
pydub>=0.25.1